refactor: move getDefaultCurrencies
