+++ apps/cowswap-frontend/src/common/modules/tradeNavigation/index.ts
@@ -8,4 +8,5 @@ export { useTradeTypeInfoFromUrl } from './hooks/useTradeTypeInfoFromUrl'
 // Util
 export { addChainIdToRoute, parameterizeTradeRoute } from './utils/parameterizeTradeRoute'
 export { parameterizeTradeSearch } from './utils/parameterizeTradeSearch'
+export { getDefaultCurrencies } from './utils/getDefaultCurrencies'
 export type { TradeSearchParams } from './utils/parameterizeTradeSearch'
