+++ apps/cowswap-frontend/src/modules/trade/types/TradeRawState.ts
@@ -1,15 +1,11 @@
-import { TokenWithLogo, USDC, WRAPPED_NATIVE_CURRENCIES as WETH } from '@cowprotocol/common-const'
 import { isEvmChain, OrderKind, SupportedChainId } from '@cowprotocol/cow-sdk'
+
+import { getDefaultCurrencies } from 'common/modules/tradeNavigation'
 
 export interface ExtendedTradeRawState extends TradeRawState {
   readonly inputCurrencyAmount: string | null
   readonly outputCurrencyAmount: string | null
   readonly orderKind: OrderKind
-}
-
-export type TradeCurrencies = {
-  inputCurrency: TokenWithLogo | null
-  outputCurrency: TokenWithLogo | null
 }
 
 export interface TradeRawState {
@@ -19,13 +15,6 @@ export interface TradeRawState {
   readonly outputCurrencyId: string | null
   readonly recipient?: string | null
   readonly recipientAddress?: string | null
-}
-
-export function getDefaultCurrencies(chainId: SupportedChainId | null): TradeCurrencies {
-  return {
-    inputCurrency: chainId ? WETH[chainId] || null : null,
-    outputCurrency: chainId ? USDC[chainId] || null : null,
-  }
 }
 
 export function getDefaultTradeRawState(chainId: SupportedChainId | null): TradeRawState {
